Fixes: Cast/Kill threads weren't getting deleted properly when the monster gets overwhelmed.  Auto wield is fixed.  Ring wearing is fixed.  Kobold aura check was too stringent.
The problem was that .notify() called from a loop in MudReader was causing the same list to have an object removed.  For example, KillThread.notify() would get called when a monster was killed, so the KillThread would destroy itself, which would modify the list currently being iterated in MudReader.  The solution was to add 'unregistered' to BotReactions, a field which can be checked when MudReader is done looping, and the list is modified after the loop.

Auto wield and ring weild make use of the now-working BotReactions.

The kobolds get unsafe when you're 'blue' so it's unsafe to go at 'pale blue,' but  'dusty blue' is fine.

Wie2 had a problem in the regex.

CoolAbility and CoolAbilityThread is still obviously a work in progress.