Added ability to toggle Debug / Verbose modes
By default Debug and Verbose mode are enabled.  If you emote "defecate"
then the magentaprint is mostly disabled (exceptions to this are enabled
by passing in a second arg set to False).  Similar to this you can type
"cackle" to hide all of the telnet output except for the aforementioned
magenta.

Both of these commands work as toggles so typing the command reverses
the current state.  Enjoy!