EXP per min code
This is not working just yet but I'm checking it in commented out incase
you want to check it out / take a crack at it.  The helper function
calculate_vpm would allow you to check any numeric value compared to the
runtime in minutes (Gold, Kills, Misses, Hits, Spells Cast etc...)

Right now there is a bug in the code caused by a divide by 0.  Ideally
the code should assume at the very least that the runtime has been 1
minute and then round correctly to the nearest minute from there.  I'm
not familiar enough with python to do this quite yet tho :/