Added getLevelPath and Mage class
Since mage is really easy I added the LevelPath as an array with two
child arrays the idea is that one would take the bot to the trainer
where the bot could pause and output "train" then the next array would
take it home.  If there is way to include the pause / train in a single
array that might be better but could make our paths turn into monsters.