Buffs implemented, Added Meditate, Bug fix
Buffs have been added to the bottom of the rest_and_check_aura function.
Code for Meditate has been added that checks to see if the Title is set
to Monk.  A bug related to combat healing has been fixed (code was
copied from heal_up where it used a heal_spell variable instead of
"vig")