Updated Database Code and Tests
Instead of simply saving an Area there is now a series of business logic
checks to ensure that Areas that already exist cannot be added.  Areas
are checked using the Area title and Obvious exits supplied by the mud
reader to search for other areas that match.  In the event of a match on
a basic level a more thorough check will be done - if all checks turn up
with the match then the code will then simply return that matched record
instead of updating anything in the database.