Added empty Monster List
An empty Monster_List update has been added when entering areas that
have no monsters.  This will stop the bot from thinking the mobs from
the last area are in the new one.

Another empty of the list has been added when the text "They are not
here\." is matched.  This code is intended to get rid of a bug related
to the bot endlessly trying to attack a mob that is not there.