Added logic to remove dark areas from mapping
With the current logic in place - dark areas function like warp portals ultimately connecting nodes that shouldn't be connected.  I've added code to the MudReaderThread to remove the AreaID and LastDirection so that this wont' happen.  This code should probably be moved to the cartographer and prompt a Crawler to cast light.