Implemented Hidden AreaExits
I've added a bool to AreaExits that allow them to be identified as a
Hidden exit.  Hidden exits still have to be manually added to the
database but with this bool in place they won't be used when trying to
find the AreaID - before this update a manually added hidden exit would
cause an Area to be re-created in the database.