New: Info.
It updates character with all of the information from the info command.  I'd like to do a little bit more work on how auras are done soon.  So there's a terrible call to Aura in Info which is so things match up with the old system (AURA_PREFERRED_SCALE etc.)