SmartCombat is sort of working.
Time needs to be called on login to update the ability cooldowns.  I'm unhappy with this right now because lag screws it up pretty badly and the inheritance structure didn't work well for me.  I think I learned what a class variable is though, and how if you inherit it, it's the same variable across siblings.