Grind thread update
"go no exit" is still showing up randomly either due to lag or some
other issue so I've added some code to remove the default "pop" behavior
and have the bot retry up to 10 times.  If it can't go then it will walk
itself to the chapel and start over.