Area and Cartography efficiencies - Description is no longer checked while matching areas - Area matching uses a new query to eliminate processing and sets a flag when identical names / direction_lists are found (eg. Forest Path, [East, West] x 16+ times). - Cartography now uses new MudArea functions to discern the new area id without worrying about parsing the area name / exit list - this will help to stop uneccessary searching - DatabaseTest.py includes functions to run tests on the new Area and MudArea functions
A few notes about these changes:
When mapping an area the could would normally iterate through every area with an identical name and check for matching exits.
This appeared to be seamless but on further inspection I found there are up to 100 areas with the name "Forest Path".
To fix this I managed to write a query that given an area name and direction list it returns the exact direction list I would have generated by iterating through those 70+ areas for "Forest Path".
Along with the MudArea changes the interaction with the database should be at a very minimal level.