Added Confused bool for the no_see bug
Sometimes the bot gets stuck endlessly trying to hit something that
isn’t around and it was solved by triggering a “look” command when the
no_see regex proc’d.  This became a problem when a look was performed
after every mob kill (due to inefficient kill tracking which would be
fixed by proper CombatReactions)

The confused bool is set to True when no_see is proc’d and unset after
a successful move.  The look command is triggered on the second proc of
no_see where confused is True.