Moved chase hooks back into engage
At the end of the engage function a check is made whether or not to
chase which performs a move towards the fleeing mob.  The get_items
function checks the chase_mob before trying to get items (if a mob ran
there is no reason to get items off the ground) and finally in the
decide_which_mob_to_kill function the chase variables are unset after
the chase_mob is re-targeted.  EFFING SEAMLESS.