Inventory dict overhaul and equipment tracking
The standard {‘string’:int} storage for inventory has been replaced
with a {MudItem:MudList}.  The “keep list” becomes an array of MudItems
on init.  The equipped items are parsed when the bot looks at itself
(character is passed into inventory as a way to lookup the name and
maybe keep list / item preferences later?).

Also fixed a bug with gold coins being added to the inventory.