New: classes CombatObject and Command.
Lots of stuff in here.  The cooldowns are now done as part of the CombatObject class.  Kill and Cast are completely redone.  The cast cooldown now depends on the character class, which is very good because different classes have different cast rates.  Favourite_spell now gets set automatically.  The Ability list gets populated.  SmartCombat is working.