TelnetHandler daemon keeps connection open longer.  Go object.  "kk2" command for smart combat with lvl 2 spells.
The problem with Go is that it's faster not to wait for the command response, so it needs more work before we use it.  On the plus side, with it you can queue up a long direction list manually, but it's unfortunately too slow right now.  I'm glad I tried it though.