Flush input on exit.  Some ability/skill tweaks for bot.  Mob flee updated, chase_mob added to Go.  Death regex changed. ki fix.
I gave Go chase_mob because it should handle monster_list things and there are problems with giving that job to the combat threads (ie. mob dies if not using a combat thread).  Go and Cartography will always be close.  There's an issue with Go and doors - 'go do' makes go translate it to down.