Fixed Equipment matching.
It would miss some equipment if the server message was sent in two clumps.  Now it waits for the prompt before setting its flag.  This was also problematic, but it finally works.  I edited MudReaderThread so that if one regexes occurs many times in one clump, it can send many notifications.