Fix: Equipment command.
It was still buggy even if waiting for the prompt, since the prompt can come in between many notifications for one_equip, so the solution wasn't completed.  I outfitted MudReader with the capability to notify that all text in the buffer has been matched.  Equipment subscribes to this and only returns (sets self.waiter_flag) when the buffer has been completely processed.  I think this will solve the problem the equipment dict is missing many pieces on the end and the bot buys armour unnecessarily.

I also filled out a bit of ranger gear item data in the db.