Should kill mob update This adjusts logic around tough fights to ensure the bot meets certain criteria before engaging.
Added "should_kill_mob" check into GrindThread and overrode it in SmartGrind. It has a bit of logic to check if the mob is weak which is basically:
if (mob.level < char.level - 4)

Certain mobs get a level modifier too. If the mob is considered "tough" then the bot has to be 95% max hp and mana to engage, have a few pots and as you can see it has to check its aura and the mob still needs to be in the list. I could update this to say that the bot should be on aura but I think the pots and max stats make up for it.. hopefully this fixes deaths!

DB updated to support steel collars and masks as well.